refactor(site-transfers): migrate components to feature directory and extract TransferTable (Phase 6.1g)
- Move 'TransferRow' from shared resources to 'features/site-transfer/components'.
- Extract 'TransferTable' component to handle search, filtering, and pagination.
- Update site transfers list page to use the new feature components.
