+++ features/site-transfer/components/index.tsx
@@ -1,0 +1,2 @@
+export { TransferTable } from './transfer-table';
+export { TransferRow } from './transfer-row';
